requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
datasets==3.0.1
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.95.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
openai==0.27.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pydantic==1.10.18
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.26.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.22.0
xxhash==3.5.0
yarl==1.13.1
